Extract Internet provider tests to separate class
